chore(cleanup, fix logic): cleanup & fix remaining issues in logic

+++ src/noteCollection/noteCollections.service.ts
@@ -6,10 +6,21 @@ import { NoteCollection, Prisma } from '@prisma/client';
 export class NoteCollectionsService {
   constructor(private prisma: PrismaService) {}
 
+  async findMany(params: {
+    where: Prisma.NoteCollectionWhereInput;
+    include: Prisma.NoteCollectionInclude;
+  }): Promise<NoteCollection[]> {
+    const { where, include } = params;
+    return this.prisma.noteCollection.findMany({
+      where,
+      include,
+    });
+  }
+
   async findOne(params: {
     req: any;
     where: Prisma.NoteCollectionWhereUniqueInput;
-    include?: Prisma.NoteCollectionInclude;
+    include: Prisma.NoteCollectionInclude;
   }): Promise<NoteCollection> {
     const { req, where, include } = params;
     const noteCollection = await this.prisma.noteCollection.findUnique({
@@ -22,47 +33,44 @@ export class NoteCollectionsService {
     return noteCollection;
   }
 
-  async findMany(params: {
-    skip?: number;
-    take?: number;
-    cursor?: Prisma.NoteCollectionWhereUniqueInput;
-    where?: Prisma.NoteCollectionWhereInput;
-    orderBy?: Prisma.NoteCollectionOrderByWithRelationInput;
-    include?: Prisma.NoteCollectionInclude;
-  }): Promise<NoteCollection[]> {
-    const { skip, take, cursor, where, orderBy, include } = params;
-    return this.prisma.noteCollection.findMany({
-      skip,
-      take,
-      cursor,
-      where,
-      orderBy,
-      include,
-    });
-  }
-
-  async createNoteCollection(
-    data: Prisma.NoteCollectionCreateInput,
-  ): Promise<NoteCollection> {
+  async createNoteCollection(params: {
+    data: Prisma.NoteCollectionCreateInput;
+  }): Promise<NoteCollection> {
+    const { data } = params;
     return this.prisma.noteCollection.create({
       data,
     });
   }
 
   async updateNoteCollection(params: {
+    req: any;
     where: Prisma.NoteCollectionWhereUniqueInput;
     data: Prisma.NoteCollectionUpdateInput;
   }): Promise<NoteCollection> {
-    const { where, data } = params;
-    return this.prisma.noteCollection.update({
+    const { req, where, data } = params;
+    const noteCollection = await this.prisma.noteCollection.findUnique({
       where,
+    });
+    if (!noteCollection || noteCollection.userId !== req.user.id) {
+      throw new NotFoundException('NoteCollection not found');
+    }
+    return this.prisma.noteCollection.update({
       data,
+      where,
     });
   }
 
-  async deleteNoteCollection(
-    where: Prisma.NoteCollectionWhereUniqueInput,
-  ): Promise<NoteCollection> {
+  async deleteNoteCollection(params: {
+    req: any;
+    where: Prisma.NoteCollectionWhereUniqueInput;
+  }): Promise<NoteCollection> {
+    const { req, where } = params;
+    const noteCollection = await this.prisma.noteCollection.findUnique({
+      where,
+    });
+    if (!noteCollection || noteCollection.userId !== req.user.id) {
+      throw new NotFoundException('NoteCollection not found');
+    }
     return this.prisma.noteCollection.delete({
       where,
     });
