+++ src/notes/notes.service.ts
@@ -6,10 +6,21 @@ import { Note, Prisma } from '@prisma/client';
 export class NotesService {
   constructor(private prisma: PrismaService) {}
 
+  async findMany(params: {
+    where: Prisma.NoteWhereInput;
+    include: Prisma.NoteInclude;
+  }): Promise<Note[]> {
+    const { where, include } = params;
+    return this.prisma.note.findMany({
+      where,
+      include,
+    });
+  }
+
   async findOne(params: {
     req: any;
     where: Prisma.NoteWhereUniqueInput;
-    include?: Prisma.NoteInclude;
+    include: Prisma.NoteInclude;
   }): Promise<Note> {
     const { req, where, include } = params;
     const note = await this.prisma.note.findUnique({
@@ -22,43 +33,56 @@ export class NotesService {
     return note;
   }
 
-  async findMany(params: {
-    skip?: number;
-    take?: number;
-    cursor?: Prisma.NoteWhereUniqueInput;
-    where?: Prisma.NoteWhereInput;
-    orderBy?: Prisma.NoteOrderByWithRelationInput;
-    include?: Prisma.NoteInclude;
-  }): Promise<Note[]> {
-    const { skip, take, cursor, where, orderBy, include } = params;
-    return this.prisma.note.findMany({
-      skip,
-      take,
-      cursor,
+  async createNote(params: {
+    req: any;
+    where: Prisma.NoteWhereUniqueInput;
+    data: Prisma.NoteCreateInput;
+  }): Promise<Note> {
+    const { req, where, data } = params;
+    const noteCollection = await this.prisma.noteCollection.findUnique({
       where,
-      orderBy,
-      include,
     });
-  }
-
-  async createNote(data: Prisma.NoteCreateInput): Promise<Note> {
+    if (!noteCollection || noteCollection.userId !== req.user.id) {
+      throw new NotFoundException('NoteCollection not found');
+    }
     return this.prisma.note.create({
       data,
     });
   }
 
   async updateNote(params: {
+    req: any;
     where: Prisma.NoteWhereUniqueInput;
     data: Prisma.NoteUpdateInput;
+    include: Prisma.NoteInclude;
   }): Promise<Note> {
-    const { where, data } = params;
+    const { req, where, data, include } = params;
+    const note = await this.prisma.note.findUnique({
+      where,
+      include,
+    });
+    if (!note || note.NoteCollection.userId !== req.user.id) {
+      throw new NotFoundException('Note not found');
+    }
     return this.prisma.note.update({
       where,
       data,
     });
   }
 
-  async deleteNote(where: Prisma.NoteWhereUniqueInput): Promise<Note> {
+  async deleteNote(params: {
+    req: any;
+    where: Prisma.NoteWhereUniqueInput;
+    include: Prisma.NoteInclude;
+  }): Promise<Note> {
+    const { req, where, include } = params;
+    const note = await this.prisma.note.findUnique({
+      where,
+      include,
+    });
+    if (!note || note.NoteCollection.userId !== req.user.id) {
+      throw new NotFoundException('Note not found');
+    }
     return this.prisma.note.delete({
       where,
     });
